+++ CHANGELOG.md
@@ -16,6 +16,9 @@ Now you are able to use `micrometer-registry-influx`, for more info [see our doc
 ### Features
 1. [#223](https://github.com/influxdata/influxdb-client-java/pull/223): Exponential random backoff retry strategy
 
+### Bug Fixes
+1. [#???](https://github.com/influxdata/influxdb-client-java/pull/???): Connection URL with custom base path
+
 ## 2.2.0 [2021-04-30]
 
 ### Breaking Changes